requirements.txt
certifi==2022.6.15
click==8.0.4
colorama==0.4.4
dnspython==1.16.0
email-validator==1.1.3
Flask==2.0.3
Flask-WTF==0.15.1
gunicorn==19.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
neo4j==5.2.1
pip==22.2.2
python-dotenv==0.21.0
pytz==2022.6
setuptools==65.5.0
Werkzeug==2.0.3
wheel==0.37.1
wincertstore==0.2
WTForms==2.3.3
